// src/app/models/detail-column.model.ts
export enum DetailColumnID {
    uic = 0,
    unitName = 1,
    utc = 2,
    location = 3,
    metl = 4,
    metlProj = 5,
    metlAsOf = 6,
    metlAge = 7,
    assigned = 8,
    majcom = 9,
    cRating = 10,
    cRatingProj = 11,
    cRatingAsOf = 12,
    cRatingAge = 13,
    pRating = 14,
    sRating = 15,
    rRating = 16,
    tRating = 17,
    activity = 18,
    metlPrev = 19,
    cRatingPrev = 20,
    pctef = 21,
}